r/aws_iam_role: Use 'roleCreateTags'.
